img 返回值更保险的意外处理

+++ light_recognition.h
@@ -36,7 +36,7 @@ typedef struct {
 } Point;
 
 typedef struct {
-    int cx, cy, radius;
+    float cx, cy, radius;
 } Circle;
 
 void cvHandle(void);
